add user reducer
